+++ plugins/11_TheWarWithin/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Undermine
   * Add Delve pets to tooltips
   * Add Undermine Undershirt secret reward
+  * Add Authentic Undermine Clam Chowder secret recipe
   * Add new vendor - Bilgewater quatermaster
   * Add new vendor - Steamwheedle quatermaster
   * Add new vendor - Blackwater quatermaster